Update API wrappers

+++ api/oneinch/health.go
@@ -7,7 +7,7 @@ import (
 )
 
 // GetRouterHealthStatus checks the status of the 1inch router
-func GetRouterHealthStatus(chainID int) error {
+func GetRouterHealthStatus(chainID uint64) error {
 	client := &http.Client{}
 	req, err := http.NewRequest("GET", ApiBaseUrl+"/"+ApiVersion+"/"+fmt.Sprint(chainID)+string(HealthcheckEndpoint), nil)
 	if err != nil {
